code review refactor

+++ kasper-security/src/main/java/com/viadeo/kasper/security/authz/commands/handlers/role/RemoveRoleFromGroupCommandHandler.java
@@ -7,6 +7,7 @@
 package com.viadeo.kasper.security.authz.commands.handlers.role;
 
 import com.google.common.base.Optional;
+import com.viadeo.kasper.CoreReasonCode;
 import com.viadeo.kasper.KasperID;
 import com.viadeo.kasper.cqrs.command.CommandResponse;
 import com.viadeo.kasper.cqrs.command.EntityCommandHandler;
@@ -18,36 +19,45 @@ import com.viadeo.kasper.security.authz.commands.role.RemoveRoleFromGroupCommand
 import com.viadeo.kasper.security.authz.entities.actor.Group;
 import com.viadeo.kasper.security.authz.entities.permission.impl.Role;
 import com.viadeo.kasper.security.authz.entities.relations.Group_has_Role;
+import org.axonframework.repository.AggregateNotFoundException;
 
 @XKasperCommandHandler(domain = Authorization.class, description = "Remove a permission from a group for authorizations")
 public class RemoveRoleFromGroupCommandHandler extends EntityCommandHandler<RemoveRoleFromGroupCommand, Group_has_Role> {
 
     @Override
     public CommandResponse handle(final KasperCommandMessage<RemoveRoleFromGroupCommand> message) throws Exception {
-        final Group_has_Role groupHasRole = new Group_has_Role(
-                this.getGroup(message.getCommand().getGroupId()),
-                this.getRole(message.getCommand().getRoleId())
-        );
-        groupHasRole.delete();
-        return CommandResponse.ok();
+        final Optional<Group> group = this.getGroup(message.getCommand().getGroupId());
+        final Optional<Role> role = this.getRole(message.getCommand().getRoleId());
+        if (role.isPresent() && group.isPresent()) {
+            final Group_has_Role groupHasRole = new Group_has_Role(group.get(), role.get());
+            groupHasRole.delete();
+            return CommandResponse.ok();
+        } else {
+            return CommandResponse.error(CoreReasonCode.INVALID_INPUT);
+        }
     }
 
-    public Role getRole(final KasperID id) {
-        Role role = null;
+    public Optional<Role> getRole(final KasperID id) {
         final Optional<ClientRepository<Role>> roleRepositoryOpt = this.getRepositoryOf(Role.class);
         if (roleRepositoryOpt.isPresent()) {
-            role = roleRepositoryOpt.get().business().get(id);
+            try {
+                return Optional.of(roleRepositoryOpt.get().business().get(id));
+            } catch (AggregateNotFoundException e) {
+                return Optional.absent();
+            }
         }
-        return role;
+        return Optional.absent();
     }
 
-    public Group getGroup(final KasperID id) {
-        Group group = null;
+    public Optional<Group> getGroup(final KasperID id) {
         final Optional<ClientRepository<Group>> groupRepositoryOpt = this.getRepositoryOf(Group.class);
         if (groupRepositoryOpt.isPresent()) {
-            group = groupRepositoryOpt.get().business().get(id);
+            try {
+                return Optional.of(groupRepositoryOpt.get().business().get(id));
+            } catch (AggregateNotFoundException e) {
+                return Optional.absent();
+            }
         }
-        return group;
+        return Optional.absent();
     }
-
 }
