+++ kasper-security/src/main/java/com/viadeo/kasper/security/authz/commands/handlers/role/RemoveRoleFromUserCommandHandler.java
@@ -7,6 +7,7 @@
 package com.viadeo.kasper.security.authz.commands.handlers.role;
 
 import com.google.common.base.Optional;
+import com.viadeo.kasper.CoreReasonCode;
 import com.viadeo.kasper.KasperID;
 import com.viadeo.kasper.cqrs.command.CommandResponse;
 import com.viadeo.kasper.cqrs.command.EntityCommandHandler;
@@ -18,36 +19,46 @@ import com.viadeo.kasper.security.authz.commands.role.RemoveRoleFromUserCommand;
 import com.viadeo.kasper.security.authz.entities.actor.User;
 import com.viadeo.kasper.security.authz.entities.permission.impl.Role;
 import com.viadeo.kasper.security.authz.entities.relations.User_has_Role;
+import org.axonframework.repository.AggregateNotFoundException;
 
 @XKasperCommandHandler(domain = Authorization.class, description = "Remove a permission from a group for authorizations")
 public class RemoveRoleFromUserCommandHandler extends EntityCommandHandler<RemoveRoleFromUserCommand, User_has_Role> {
 
     @Override
     public CommandResponse handle(final KasperCommandMessage<RemoveRoleFromUserCommand> message) throws Exception {
-        final User_has_Role userHasRole = new User_has_Role(
-            this.getUser(message.getCommand().getUserId()),
-            this.getRole(message.getCommand().getRoleId())
-        );
-        userHasRole.delete();
-        return CommandResponse.ok();
+        final Optional<Role> role = this.getRole(message.getCommand().getRoleId());
+        final Optional<User> user = this.getUser(message.getCommand().getUserId());
+        if (role.isPresent() && user.isPresent()) {
+            final User_has_Role userHasRole = new User_has_Role(user.get(), role.get());
+            userHasRole.delete();
+            return CommandResponse.ok();
+        } else {
+            return CommandResponse.error(CoreReasonCode.INVALID_INPUT);
+        }
     }
 
-    public Role getRole(final KasperID id) {
-        Role role = null;
+    public Optional<Role> getRole(final KasperID id) {
         final Optional<ClientRepository<Role>> roleRepositoryOpt = this.getRepositoryOf(Role.class);
         if (roleRepositoryOpt.isPresent()) {
-            role = roleRepositoryOpt.get().business().get(id);
+            try {
+                return Optional.of(roleRepositoryOpt.get().business().get(id));
+            } catch (AggregateNotFoundException e) {
+                return Optional.absent();
+            }
         }
-        return role;
+        return Optional.absent();
     }
 
-    public User getUser(final KasperID id) {
-        User user = null;
+    public Optional<User> getUser(final KasperID id) {
         final Optional<ClientRepository<User>> userRepositoryOpt = this.getRepositoryOf(User.class);
         if (userRepositoryOpt.isPresent()) {
-            user = userRepositoryOpt.get().business().get(id);
+            try {
+                return Optional.of(userRepositoryOpt.get().business().get(id));
+            } catch (AggregateNotFoundException e) {
+                return Optional.absent();
+            }
         }
-        return user;
+        return Optional.absent();
     }
 
 }
